data_pipeline/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
langchain-community
langchain-text-splitters
langchain-ollama
langchain-chroma==0.2.0
chromadb==0.5.3
psycopg2-binary
